Initial Graphing Implementation and Cleanup
Actual graphing has sort of begun being implemented. I moved the main method to Graph as well, which hopefully makes some of the calls clearer. In the future, we might want to move it to somewhere else. This is currently broken at drawGraph and possibly at its call as well.