Dispatch
Tried making a new class from which to dispatch stuff to avoid most of the issues with synchronization. Honestly, IDK if this is gong to help at all. I linked a Stack thread about how we might fix it. If we can't get it running soon, we should talk to Mr. Duran about what's the best way to implement it.