It's Alive!
The project works again, perfect! I'll probably move this back to the GUI class once I can fix the performance problem with rounding (which is going to take minutes if not hours currently just to accurately fill the xValues, God forbid the yValues). Let's talk about reducing precision requirements for most of the graph and then finding a way to identify key intervals on which it needs to be increased.