yValues Filling
yValues are currently filling, only issue is performance again. We'll have to look into what the most efficient way to do this is, it might involve modifying mathEvaluator to have a way to iterate through various xValues, thereby hopefully skipping a lot of its logic which is probably what's sucking performance.