Reduced Precision Functioning.
Reduced the precision to every 10^-2 for general use to make performance tolerable. Precision can be increased in certain specific intervals if need be. All points are being filled accurately (barring floating point arithmetic artifacts) but the drawing is currently only working for the x and y axes. On to the next issue.