QoL Changes
Changed the precision to 10^-4, which seems to have acceptable performance, but also reduces the dashing that existed with functions like tan(x) at 10^-2. Currently there are some issues with graphing complex polynomial functions, but this could just be an issue with MathEvaluator. Now I'm going to move on to derivatives for now.