Drawing Key Points
All key points are now automatically drawn. Zeroes works perfectly Kevin.The small inaccuracies are because fillOval() is defined for ints, we could possibly improve this by switching to a significantly more complex method, but it's good for now. We should probably add a text display of the values, and possibly color coding.