Midway through Graphics Fix
Midway through Graphics Fix. I'm leaving off here for now because it's going to take major method design changes to complete it, and I want to leave the project functional (i.e. I need at least half a period to do it). This solution is slightly janky, but should still work with a mild decrase in graphics performance.